+++ tui/src/ui/components/vendored/mod.rs
@@ -1,1 +1,2 @@
 pub mod tui_realm_stdlib_input;
+pub mod tui_realm_stdlib_progressbar;